fix(tss): pick correct snapshot to check locking period
In the current model, we generate a new snapshot for every key. Therefore, it can happen that a new snapshot is created (e.g. for a key on another chain) before the previous key was assigned to be a master key. Instead checking if the locking period has passed for the most current snapshot, we now need to check against the snapshot that created the original key.