feat(nexus): register assets (#322)
This pull requests augments the nexus keeper with the capability of receiving information about the assets that are supported by each chain through the RegisterAssert and HasRegisteredAsset functions in the keeper. The intention is to prevent linking addresses to exchange assets that are not supported by the recipient's chain.

In addition, the e2e unit tests are augmented to also cover the act of deploying the token contract, a new query to fetch a token address by symbol was added to the ethereum module, and the event emission for a verification in bitcoin/ethereum are augmented with more attributes.