build: upgrade cosmos-sdk to 0.42.4 (#399)
First step to fully migrate axelar-core. The dependency is changed to cosmos-sdk 0.42.4, but wherever possible we still use the old handler/querier structure, use legacy REST APIs and the legacy amino codec. The rest of the migration can now be done in small chunks instead of one big breaking change