refactor(vald): replace event pipeline
We suspect we dropped events when listening to events with large throughput (keygen/sign). To avoid this, we change the event listener to only listen for new block headers and then go fetch the full block result. Then we can read all events from that block and forward it to the corresponding listeners in-process instead via websocket. This makes the architecture more robust allows resync and fault-tolerance features in the future.