fix(evm): stop copying bytecode during confirmation vote (#559)
When a chain becomes confirmed we used to copy several smart contracts from one key in the kvstore to the other. This added unnecessary gas costs to the vote that ended finalizing the poll and was difficult to calculate up front.

This commit immediately stores the smart contracts under the correct key to remove this issue. Theoretically, a malicious client could spam the blockchain with chain parameters that never get confirmed. However, these parameters behave the same as any other value stored on the blockchain, the user has to pay for them and they do not interfere with correct chains, so no special treatment is needed.