fix(evm): obtain snapshot counter and revote locking period without Ethereum chain (#634)
The confirm-chain command kicks off a poll for registering a chain. When doing this, the revote locking period and snapshot counter were obtained from a pre-existing Ethereum chain. This commit changes how these values are obtained so that an Ethereum chain is not needed. This is done by 1) fetching the snapshot counter from the latest counter rather than the snapshot associated to the Ethereum master key, and 2) using the value present on the params json supplied during add-chain