feat(evm): bytecodes and signed tx queries (#635)
In preparation for the removal of the RPCs from the EVM module, this PR introduces new queries to obtain a signed tx and the bytecodes of the contracts. Once this is merged into main, the RPCs can be removed, since we can use the information provided by these new queries to proceed with a ceremony using external means.