feat(tss): queries for active old keys (#834)
We should be able to retrieve the old rotated keys that are still active. This is particularly important to allow validators to figure out which keys they need to recover their shares from. To this end, this PR introduces the tss queries `active-old-keys` and `active-old-keys-by-validator`. The former enables us to retrieve the these keys by chain & role, whereas the later allows us to obtain the the old active keys that some validator has shares from.