fix(tss): remove signing sanity checks from end blocker (#896)
Performing so many sanity checks for signing at the end blocker is dangerous. If there is an error, we could end up on an invalid state. This happens because the state is not rolled back at the end blocker, like it happens for transactions. To mitigate such scenario, this PR moves all sanity checks to the `StartSign` keeper function. Signing is then enqueued if and only if all sanity checks pass. `StartSign` also replaces `EnqueueSign`.