refactor(tss): implement router and handlers (#898)
This PR implements a mechanism for routing completed signatures to the modules that requested them. The functionality from the evm/bitcoin end blockers have also been moved to specific module handlers that are passed to the tss module during initialization.

Nonetheless, this change does not solve the problem of reaching an invalid state due to the inability of rolling back the state, because errors returned by the handlers are logged instead of being replayed. Further refactoring will be required in future PRs to properly adapt the end blocker's code to the new mechanism.