refactor(multisig): replace panic with error
Panics should be reserved for unexpected behaviour due to implementation errors. In this case, it is sensible to expect the signature handler to return an error when handling a completed signature, so it is better to deal with that error instead of panicking.
This does not change the behaviour of the endblocker, both errors and panics are caught by the cached execution environment.