fix(axelarnet): remove axelarnet module address from bank keeper blocked list (#1835)
The module address is not allowed to send IBC transfers (which the module relies on) as long as it is on the bank keeper's blocked list. On the other hand, the axelarnet module does not need to keep the invariant `actual module balance == expected module balance`, since we never track that balance outside the bank keeper to begin with.