refactor(axelarnet)!: prepend version number
Previously, the EVM-to-Cosmos translation required Solidity ABI-encoded payloads to include the version number. This approach was inconvenient for CosmWasm contracts in Cosmos-to-Cosmos flows, as it necessitated ABI encoding.

This PR improves the version handling by prepending a 4-byte version number directly to the payload, eliminating the need for ABI encoding in CosmWasm contracts.