fix: correctly override the handler function when registering services
Previously, the loop that iterated over all handler functions of a server copied the handler, so any changes made to it where not persisted. Now, handlers are correctly wrapped, so they will always return an sdkerrors.Error, even if the handler itself doesn't specify any. This is important for vald, so it can ignore errors that happen during message execution.